requirements.txt
appdirs==1.4.4
asgiref==3.2.10
boto3==1.14.51
botocore==1.17.51
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0
django-appconf==1.0.4
django-cleanup==5.0.0
django-compressor==2.4
django-crispy-forms==1.9.2
django-extensions==3.0.5
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-postman==3.6.2
django-recaptcha==2.0.6
django-sass-processor==0.8.1
django-storages==1.9.1
django-taggit==1.3.0
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jsonfield==3.1.0
libsass==0.20.1
ndg-httpsclient==0.5.1
Pillow==7.2.0
psycopg2==2.8.5
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
s3transfer==0.3.3
scss==0.8.73
sendgrid==6.4.6
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
swapper==1.1.2.post1
urllib3==1.25.10
virtualenv==20.0.31
whitenoise==5.2.0
